Upgraded cfbenchmarks to v3 framework
