+++ .changeset/pink-flies-approve.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/cfbenchmarks-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
