+++ packages/core/legos/tsconfig.json
@@ -104,9 +104,6 @@
       "path": "../../sources/cfbenchmarks"
     },
     {
-      "path": "../../sources/cfbenchmarks-test"
-    },
-    {
       "path": "../../sources/chain-reserve-wallet"
     },
     {
