--- packages/sources/cfbenchmarks-test/CHANGELOG.md
@@ -1,126 +1,0 @@
-# @chainlink/cfbenchmarks-test-adapter
-
-## 1.4.4
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.4.3
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.4.2
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.4.1
-
-### Patch Changes
-
-- c35b8ec32: Updated custom router to default to websocket for secondary endpoints
-
-## 1.4.0
-
-### Minor Changes
-
-- 490307c73: Change default crypto transport to WS
-
-## 1.3.4
-
-### Patch Changes
-
-- 62fa38c0f: Fixed rest transports
-
-## 1.3.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.3.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.3.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.3.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.2.0
-
-### Minor Changes
-
-- 1b221edfe: Added BIRC endpoint to cfbenchmarks-test
-
-## 1.1.2
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Removed WS_ENABLED env var. Upgraded typescript to version 4.9.5
-
-## 1.1.1
-
-### Patch Changes
-
-- c152ed06e: Bumped v3 framework version
-
-## 1.1.0
-
-### Minor Changes
-
-- 70a7b9abd: Add missing rest transport in cfbenchmark v3 EA
-
-## 1.0.5
-
-### Patch Changes
-
-- b98296089: Bumped v3 framework version
-
-## 1.0.4
-
-### Patch Changes
-
-- 6d57b64a5: Tick framework version
-
-## 1.0.3
-
-### Patch Changes
-
-- f850f4831: Bump framework version
-- 917400c7b: Bump framework version
-
-## 1.0.2
-
-### Patch Changes
-
-- 6abb413ab: bump framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- 3f5b31e08: Updated env vars to standard naming
-
-## 1.0.0
-
-### Major Changes
-
-- 07781d25d: Initial release
-
-### Patch Changes
-
-- 81f5e595c: Fixed overrides and caching issues
