--- packages/sources/cfbenchmarks-test/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@chainlink/cfbenchmarks-test-adapter",
-  "version": "1.4.4",
-  "description": "Chainlink cfbenchmarks-test adapter.",
-  "keywords": [
-    "Chainlink",
-    "LINK",
-    "blockchain",
-    "oracle",
-    "cfbenchmarks-test"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "url": "https://github.com/smartcontractkit/external-adapters-js",
-    "type": "git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "tsc -b",
-    "server": "node -e 'require(\"./index.js\").server()'",
-    "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
-    "start": "yarn server:dist"
-  },
-  "dependencies": {
-    "@chainlink/external-adapter-framework": "0.28.2",
-    "tslib": "2.4.1"
-  },
-  "devDependencies": {
-    "@types/jest": "27.5.2",
-    "@types/node": "16.11.51",
-    "@types/supertest": "2.0.12",
-    "mock-socket": "9.1.5",
-    "nock": "13.2.9",
-    "supertest": "6.2.4",
-    "typescript": "4.9.5"
-  }
-}
