--- packages/sources/cfbenchmarks-test/src/endpoint/index.ts
@@ -1,2 +1,0 @@
-export { endpoint as crypto } from './common/crypto'
-export { endpoint as birc } from './rest/birc'
