--- packages/sources/cfbenchmarks-test/test-payload.json
@@ -1,7 +1,0 @@
-{
-    "requests": [{
-         "index": "DEFI_KXBTUSD"
-    }]
-}
-
-
