--- packages/sources/cfbenchmarks/schemas/env.json
@@ -1,36 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/cfbenchmarks-adapter.json",
-  "title": "Chainlink External Adapter for Cfbenchmarks",
-  "required": ["API_USERNAME", "API_PASSWORD"],
-  "type": "object",
-  "properties": {
-    "API_USERNAME": {
-      "type": "string",
-      "required": true
-    },
-    "API_PASSWORD": {
-      "type": "string",
-      "required": true
-    },
-    "API_SECONDARY": {
-      "type": "boolean",
-      "required": false,
-      "default": "false"
-    },
-    "BIRC_RETRY": {
-      "type": "number",
-      "required": false,
-      "default": 3
-    },
-    "BIRC_RETRY_DELAY_MS": {
-      "type": "number",
-      "required": false,
-      "default": 1000
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
