+++ packages/sources/cfbenchmarks/src/endpoint/index.ts
@@ -1,7 +1,2 @@
-import type { TInputParameters as ValuesInputParameters } from './values'
-import type { TInputParameters as BircInputParameters } from './birc'
-
-export type TInputParameters = ValuesInputParameters | BircInputParameters
-
-export * as values from './values'
-export * as birc from './birc'
+export { endpoint as crypto } from './common/crypto'
+export { endpoint as birc } from './rest/birc'
