+++ packages/sources/cfbenchmarks/tsconfig.json
@@ -4,7 +4,5 @@
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": ["src/**/*"],
-  "exclude": ["dist", "**/*.spec.ts", "**/*.test.ts"],
-  "references": [{ "path": "../../core/test-helpers" }, { "path": "../../core/bootstrap" }]
+  "include": ["src/**/*", "src/**/*.json"]
 }
